chore(notifications): typos

+++ src/server/mail/alias/destroy.ts
@@ -1,36 +1,31 @@
 import 'server-only'
-import { ForwardMailAlias, MailAlias, RawAddressMailAlias } from '@prisma/client';
+import { MailAlias } from '@prisma/client';
 import { prismaCall } from '../../prismaCall';
 import prisma from '@/prisma';
-import { CreateMailAliasForwardRelationTypes, createMailAliasForwardRelationValidation, destoryMailAliasValidation } from './validation';
+import { destoryMailAliasValidation } from './validation';
+import { ServerError } from '@/server/error';
 
 
 export async function destroyMailAliasById(id: number): Promise<MailAlias> {
     const parse = destoryMailAliasValidation.detailedValidate({ id })
 
+    const connectedToNotificationChannel = await prismaCall(() => prisma.mailAlias.findUniqueOrThrow({
+        where: {
+            id: parse.id,
+        },
+        select: {
+            notificationChannel: true,
+        }
+    }))
+
+    if (connectedToNotificationChannel.notificationChannel.length) {
+        throw new ServerError("BAD PARAMETERS", "Cannot delete a mailAlias that is connected to a notification channel.")
+    }
+
+
     return await prismaCall(() => prisma.mailAlias.delete({
         where: {
             id: parse.id,
         },
-    }))
-}
-
-export async function destroyMailAliasRawAddress(id: number): Promise<RawAddressMailAlias> {
-    const parse = destoryMailAliasValidation.detailedValidate({ id })
-
-    return await prismaCall(() => prisma.rawAddressMailAlias.delete({
-        where: {
-            id: parse.id,
-        }
-    }))
-}
-
-export async function destroyMailAliasForward(rawdata: CreateMailAliasForwardRelationTypes['Detailed']): Promise<ForwardMailAlias> {
-    const parse = createMailAliasForwardRelationValidation.detailedValidate(rawdata)
-
-    return await prismaCall(() => prisma.forwardMailAlias.delete({
-        where: {
-            sourceId_drainId: parse
-        }
     }))
 }
